Validate import-clr annotations against the real CLR member
Annotated (import-clr ...) declarations were trusted blindly: a wrong
type annotation silently propagated and caused cascading, misleading
errors downstream. Cross-check each annotated import against the CLR
member it binds during type inference and fail compilation on a mismatch.

The check is assignability-aware (params contravariant, return covariant),
so legitimate widenings pass — ImmutableList->IEnumerable, omitted
optional/params parameters, delegate coercion, Object widening — while
real lies are caught: wrong arity, mismatched concrete types, and enum
members declared as Int. Resolution failures and ambiguous overloads are
skipped rather than guessed, so the validator never emits a false error.

Also make the stdlib honest: register a built-in `Seq` alias for
IEnumerable<T> and re-annotate Dictionary .Keys/.Values (hash, mutable
hash, concurrent dictionary) to return (Seq ...) instead of lying about
TreeList; convert http.zs response-status-code from the HttpStatusCode
enum to Int via Convert.ToInt32 instead of declaring the enum as Int.

+++ tests/ZScheme.Compiler.Tests/Integration/EndToEndTests.cs
@@ -1831,12 +1831,12 @@ public class EndToEndTests
             @"(module test)
 (import-clr
   [stream-length System.IO.Stream.Length
-    :instance-property : (System.IO.Stream -> Int)])
+    :instance-property : (System.IO.Stream -> Long)])
 
-(define (get-length [s : System.IO.Stream]) : Int
+(define (get-length [s : System.IO.Stream]) : Long
   (stream-length s))
 
-(define (test) : Int
+(define (test) : Long
   (get-length (new System.IO.MemoryStream)))";
         var cs = Compile(source);
         Assert.Contains(".Length", cs);
